Updated onSubmit to be props dependable
